+++ packages/components/src/Icon/Icon.test.tsx
@@ -26,8 +26,12 @@
 
 import 'jest-styled-components'
 import React from 'react'
-import { assertSnapshot } from '@looker/components-test-utils'
+import { assertSnapshot, renderWithTheme } from '@looker/components-test-utils'
 import { Icon } from './Icon'
+
+test('Icon default', () => {
+  assertSnapshot(<Icon name="Plus" />)
+})
 
 test('Icon with styled system size', () => {
   assertSnapshot(<Icon name="Plus" size="large" />)
@@ -39,4 +43,11 @@ test('Icon with styled system size', () => {
 
 test('Icon with styled system size', () => {
   assertSnapshot(<Icon name="Plus" size="1rem" />)
+})
+
+test('Icon supports DOM attributes', () => {
+  const { findByLabelText } = renderWithTheme(
+    <Icon name="Plus" aria-label="Add" />
+  )
+  expect(findByLabelText('Add')).toBeTruthy()
 })