+++ storybook/.storybook/preview.js
@@ -1,0 +1,4 @@
+import { addDecorator } from '@storybook/react'
+import componentsDecorator from './componentsDecorator'
+
+addDecorator(componentsDecorator)