+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `ButtonGroup` space between rows when wrapping
+- `InputChips` separates chips by newline when pasting
 
 ## [0.8.6]