+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Icon accepts artwork as props.
 - `Fieldset` supports an accordion mode via an `accordion` prop
 
 ### Changed